cicd_agent/requirements.txt
langgraph>=0.0.50
langchain-core>=0.1.0
pyyaml>=6.0
